Autopia
Autopia is a free family Typeface designed by Antoine Gelgon. This
project was initiated during the Summer School of OSP in August 2013.
The particularity of this fonts is in the process of developing its
shape. The skeleton was generated through Autotrace Program (a program
that converts bitmap to vectorgraphics). Calligraphic forms were drawn
with Python Fontforge’s code.